Restructured Plugin
Refactored plugin for better organization and maintainability. The code that was previously in a single file (index.js) has been split into multiple files.

- Added VoiceActivity for voice chats
- Added Volume Control for users in voice chats
- Added Push to Talk for volume control for voicechat users
- Added Slider Controls for VoiceChat Users
- Added Userstates for VoiceChat Users
- couple other things...
